debug added

+++ ejo_wfb_stabilizer.py
@@ -14,6 +14,28 @@ import time
 import sys
 import os
 import struct
+import psutil
+
+
+def set_cpu_affinity(core_number):
+    pid = os.getpid()
+    p = psutil.Process(pid)
+
+    try:
+        # Get the list of available CPU cores
+        available_cores = list(range(psutil.cpu_count()))
+
+        # Check if the specified core_number is valid
+        if core_number not in available_cores:
+            raise ValueError(f"Invalid core number. Available cores: {available_cores}")
+
+        # Set the CPU affinity to the specified core
+        p.cpu_affinity([core_number])
+        print(f"CPU affinity set to core {core_number}")
+
+    except Exception as e:
+        print(f"Error: {e}")
+
 
 # Usage: python ejo_wfb_stabilizer.py [optional video file]
 # press "Q" to quit
@@ -31,11 +53,13 @@ processVar=0.03
 measVar=2
 
 # set to 1 to display full screen -- doesn't actually go full screen if your monitor rez is higher than stream rez which it probably is. TODO: monitor resolution detection
-showFullScreen = 1
+#showFullScreen = 1
+showFullScreen = 0
 
 # If test video plays too fast then increase this until it looks close enough. Varies with hardware. 
 # LEAVE AT 1 if streaming live video from WFB (unless you like a delay in your stream for some weird reason)
-delay_time = 1 
+#delay_time = 0
+delay_time = 1
 
 
 ######################## Region of Interest (ROI) ###############################
@@ -45,7 +69,7 @@ delay_time = 1
 # Higher FPS and lower resolution can go higher in ROI (and probably should)
 # Set showrectROI and/or showUnstabilized vars to = 1 to see the area being processed. On slower PC's 3 might be required if 720p input
 #roiDiv = 3.5
-roiDiv = 3.0
+roiDiv = 3.5
 
 # set to 1 to show the ROI rectangle 
 showrectROI = 0
@@ -81,6 +105,7 @@ SRC = 'udpsrc port=5600 buffer-size=65536 caps="application/x-rtp, media=(string
 
 qOpenHD = 'gnome-terminal -e "/home/home/qopenhd25/build-QOpenHD-Desktop_Qt_5_15_2_GCC_64bit-Debug/debug/QOpenHD"'
 
+
 # Command string with quotes
 #command_string = 'gnome-terminal -e \'/home/home/qopenhd25/build-QOpenHD-Desktop_Qt_5_15_2_GCC_64bit-Debug/debug/QOpenHD\''
 #command_string = '/home/home/qopenhd.sh transparent'
@@ -96,14 +121,75 @@ sed_commands = (
     "sed -i 's/^qopenhd_primary_video_rtp_input_port=.*/qopenhd_primary_video_rtp_input_port=5599/' /home/home/.config/OpenHD/QOpenHD.conf"
 )
 
+class PerfCounter:
+    def __init__(self, name, value):
+        self.name = name        
+        self.ttl = value                       
+        self.min = value        
+        self.max = value 
+        self.avg = value
+        self.count=1
+        self.created=time.time()
+
+    def add(self, value):
+        self.count += 1
+        if value<self.min:
+            self.min=value
+        if value>self.max:
+            self.max=value
+        self.ttl+=value
+        self.avg=self.ttl/self.count
+			
+# Create an empty dictionary
+perfs = {}
+
 # Now you can use subprocess to run the combined sed commands
 subprocess.run(sed_commands, shell=True)
+
+lastticks=time.time()
+
+showdebug=1
+currentstep=0
+procstart=time.time()
+
+def i(str, step=0):
+	global lastticks,currentstep,procstart, perfs
+	suffix=""
+	if showdebug==1 :
+		if step==1 :
+			if step in perfs and (time.time()-perfs[step].created)>1:
+				print()
+				for index, (key, value) in enumerate(perfs.items()):
+					#print(f"Counter {index + 1}: {key}")
+					print(f"{key}"+ " : " + value.name[:20].ljust(20) + " min:" + f"{value.min*1000:.1f}" 
+					+ " max:" + f"{value.max*1000:.1f}".ljust(6) + " avg:" + f"{value.avg*1000:.1f}".ljust(6)) 
+				perfs = {}				    
+
+			currentstep=1			
+			diff=time.time() - procstart * 1000
+			suffix = f" | {(time.time() - procstart) * 1000:.1f}"
+			procstart=time.time()			
+
+		if step == 0 :
+			currentstep=currentstep+1
+		global lastticks
+		elapsed=time.time()-lastticks
+		lastticks=time.time()
+		#print(f"{currentstep}"+ " : " + str + " = " + f"{elapsed*1000}") #//f"debug_step:8 : {time.time():.4f}")
+		p = PerfCounter(str, round(elapsed * 1000))
+		if currentstep in perfs:			
+			perfs[currentstep].add(elapsed)						
+		else:
+			perfs[currentstep]=PerfCounter(str,elapsed)
+		
+		#print(f"{currentstep} : {str[:20].ljust(20)} = {elapsed * 1000:.1f}" + suffix)
 
 def bring_to_foreground(process_id):
     try:
         subprocess.run(["wmctrl", "-ia", str(process_id)])
     except Exception as e:
         print(f"Error bringing window to foreground: {e}")
+
 process = None
 process_id = None 
 
@@ -114,19 +200,20 @@ process_id = None
 # bring_to_foreground(process_id) # Bring the window to the foreground
 
 # try to keep CPU up
-try:
-    fd = os.open("/dev/cpu_dma_latency", os.O_WRONLY)
-except OSError as e:
-    print(f"Error opening /dev/cpu_dma_latency: {e}")     
-try:
-    os.write(fd, struct.pack("I", 0))
-except OSError as e:
-    print(f"Error writing to /dev/cpu_dma_latency: {e}")    
-finally:
-    #os.close(fd) #keep file open
-	print(f"Error opening /dev/cpu_dma_latency: {e}")     
+# try:
+#     fd = os.open("/dev/cpu_dma_latency", os.O_WRONLY)
+# except OSError as e:
+#     print(f"Error opening /dev/cpu_dma_latency: {e}")     
+# try:
+#     #os.write(fd, struct.pack("I", 0))
+#     print(f"SKIPPED...")     
+# except OSError as e:
+#     print(f"Error writing to /dev/cpu_dma_latency: {e}")    
+# finally:
+#     #os.close(fd) #keep file open
+# 	print(f"/dev/cpu_dma_latency: Opened!")     
 
-
+#set_cpu_affinity(1)
 
 
 lk_params = dict( winSize  = (15,15),maxLevel = 3,criteria = (cv2.TERM_CRITERIA_EPS | cv2.TERM_CRITERIA_COUNT, 10, 0.03))
@@ -146,8 +233,18 @@ if len(sys.argv) == 2:
 
 video = cv2.VideoCapture(SRC)
 
-while True:
-	grab, frame = video.read()
+while True:	
+	#grab, frame = video.read()
+	i(f"Frame start",1)   #debug_step:1 : {time.time():.3f}
+	while not video.grab():
+		i("No frame grabbed, trying again...")
+
+	i(f"Grabbed ")
+
+	grab, frame = video.retrieve()
+
+	i(f"retrieved")
+
 	if grab is not True:
 		exit() 
 	res_w_orig = frame.shape[1]
@@ -175,10 +272,13 @@ while True:
 		# Not in use, save for later
 		#gfftmask = np.zeros_like(currGray)
 		#gfftmask[top_left[0]:bottom_right[0], top_left[1]:bottom_right[1]] = 255
-		
-		prevPts = cv2.goodFeaturesToTrack(prevGray,maxCorners=400,qualityLevel=0.01,minDistance=30,blockSize=3)
+		i(f"converted to gray")
+		#prevPts = cv2.goodFeaturesToTrack(prevGray,maxCorners=400,qualityLevel=0.01,minDistance=30,blockSize=3)
+		prevPts = cv2.goodFeaturesToTrack(prevGray,maxCorners=400,qualityLevel=0.1,minDistance=7,blockSize=3)
+		i(f"goodFeaturesToTrack")
 		if prevPts is not None:
 			currPts, status, err = cv2.calcOpticalFlowPyrLK(prevGray,currGray,prevPts,None,**lk_params)
+			i(f"calcOpticalFlowPyrLK")
 			assert prevPts.shape == currPts.shape
 			idx = np.where(status == 1)[0]
 			# Add orig video resolution pts to roi pts
@@ -245,19 +345,26 @@ while True:
 			if process_id != None: 
 				bring_to_foreground(process_id) # Bring the window to the foreground
 		
+		i(f"Frame ready")
 		cv2.imshow(window_name, f_stabilized)
-		
+		i(f"imshow completed")
+
 		if showUnstabilized == 1:
 			cv2.imshow("Unstabilized ROI",prevGray)
-		if cv2.waitKey(delay_time) & 0xFF == ord('q'):
+
+		#this will wait much more than 1 second and will cause CPU to go into idle state!!!	
+		#if cv2.waitKey(delay_time) & 0xFF == ord('q'):
+		if cv2.pollKey() & 0xFF == ord('q'):
 			break
-		
-		if process==None:
+
+		i(f"Cycle completed")
+
+		if process==None  and showFullScreen == 1:
 			# Start your process
 			process = subprocess.Popen(command_list)
 			#process.wait() # Wait for a moment to ensure the window is created
-			time.sleep(1)
-			process_id = process.pid # Get the process ID (PID) of the last process
+			time.sleep(1) 
+			process_id = process.pid # Get the process ID (PID) of the last process			
 			bring_to_foreground(process_id) # Bring the window to the foreground
 		
 		prevOrig = Orig
@@ -272,11 +379,14 @@ video.release()
  
 cv2.destroyAllWindows()
 
+if process is not None:
+    process.terminate()
+    # Wait for an additional 5 seconds for the process to respond to the terminate signal
+    try:
+        process.wait(timeout=1)
+    except subprocess.TimeoutExpired:
+        # If the process is still running, forcefully kill it
+        process.kill()
+else:
+    print("No qOpenHD to close!")
 
-process.terminate()
- # Wait for an additional 5 seconds for the process to respond to the terminate signal
-try:
-    process.wait(timeout=1)
-except subprocess.TimeoutExpired:
-    # If the process is still running, forcefully kill it
-    process.kill()
